pyproject.toml
[tool.poetry]
name = "aocleaderboardbot"
version = "0.1.0"
description = ""
authors = ["Jay Madden <jaymaddencox@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
lin0-length = 120
include = '\.pyi?$'
